requirements.txt
pymupdf>=1.23.0
pdfplumber>=0.10.0
openpyxl==3.1.5
requests==2.33.0
python-dotenv==1.2.0
pydantic==2.13.0
flask==3.0.0
landingai-ade==1.11.0
openai>=1.0.0
boto3>=1.34.0
openai>=1.0.0